requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-import-export==2.3.0
django-jazzmin==2.4.7
django-js-asset==1.2.2
django-model-utils==4.0.0
djangorestframework==3.11.1
drf-yasg==1.17.1
et-xmlfile==1.0.1
idna==2.10
inflection==0.5.1
isort==5.5.2
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.4
Pillow==7.2.0
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sendgrid==6.5.0
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
swagger-spec-validator==2.7.3
tablib==2.0.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
mysqlclient==2.0.3